create queue
creation of queue

install of the management tool
creation of the queue
still needs to pass credentials into management tool
need to check if the server is running to download it.

exchanges and bindings

change the path to /usr/sbin/rabbitmqadmin

checking if the core module exists

Check if the rabbitmqadmin salt module is available.

move changes from init to install